requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
customtkinter>=5.2.0
pywinstyles>=1.8
PySide6>=6.6.0
